chore(lint): rename internal/cmd/manager/debug to diagnostics
Rename the debug package to diagnostics to avoid conflicting with
the Go standard library "debug" package name. This eliminates the
need for a revive linter exclusion rule.
