+++ internal/cmd/manager/diagnostics/architectures/cmd.go
